Move Tandy sound definitions into `tandy_sound.h`
