+++ src/audio/mixer.cpp
@@ -26,6 +26,7 @@
 #include "config/setup.h"
 #include "gui/mapper.h"
 #include "hardware/audio/gus.h"
+#include "hardware/audio/tandy_sound.h"
 #include "hardware/hardware.h"
 #include "hardware/memory.h"
 #include "hardware/pic.h"
