+++ src/ints/bios.cpp
@@ -14,6 +14,7 @@
 #include "cpu/registers.h"
 #include "dos/dos_memory.h"
 #include "dosbox.h"
+#include "hardware/audio/tandy_sound.h"
 #include "hardware/hardware.h"
 #include "hardware/input/joystick.h"
 #include "hardware/input/mouse.h"
